backend/requirements.txt
Flask==2.3.3
pandas==2.1.1
openpyxl==3.1.2
python-docx==0.8.11
reportlab==4.0.4
python-dateutil==2.8.2
gunicorn==21.2.0
python-dotenv==1.0.0
